给所有实体类添加公有继承类
为了解决id，创建修改时间，添加了baseEntity类为所有实体类的父类。但是带来了一些bug还有部分没有修复

+++ xyblog-backEnd/xyblog-admin/src/main/java/cn/xinyuan/blog/admin/article/service/IBlogArticleInfoService.java
@@ -24,5 +24,5 @@ public interface IBlogArticleInfoService extends IService<BlogArticleInfo> {
 
     int updateArticle(ArticleContentDTO article);
 
-    boolean checkByCategory(Integer id);
+    boolean checkByCategory(Long id);
 }
