+++ xyblog-backEnd/xyblog-common/src/main/java/cn/xinyuan/blog/common/base/AbstractController.java
@@ -14,7 +14,7 @@ public class AbstractController {
         return (SysUser) SecurityUtils.getSubject().getPrincipal();
     }
 
-    protected Integer getUserId(){
+    protected Long getUserId(){
         return getUser().getId();
     }
 }
