+++ xyblog-backEnd/xyblog-common/src/main/java/cn/xinyuan/blog/common/constants/SysConstants.java
@@ -12,6 +12,6 @@ public class SysConstants {
     /**
      * 超级管理员账号Id
      */
-    public static Integer SUPER_ADMIN=1;
+    public static Long SUPER_ADMIN= 1L;
 
 }
